requirements.txt
google-genai>=1.0.0
python-chess>=1.10.0
Pillow>=10.0.0
openai>=1.0.0
